# server/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Record_string.any_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"UserRole": {
				"type": "string",
				"enum": [
					"admin",
					"user"
				]
			},
			"SafeUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"settings": {
						"$ref": "#/components/schemas/Record_string.any_"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"email",
					"name",
					"role",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDTO": {
				"properties": {
					"email": {
						"type": "string"
					},
					"passwordHash": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"settings": {
						"$ref": "#/components/schemas/Record_string.any_"
					}
				},
				"required": [
					"email",
					"passwordHash",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_UserDTO_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"passwordHash": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string"
					},
					"settings": {
						"$ref": "#/components/schemas/Record_string.any_"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateUserDTO": {
				"$ref": "#/components/schemas/Partial_UserDTO_"
			},
			"Difficulty": {
				"type": "string",
				"enum": [
					"easy",
					"medium",
					"hard"
				]
			},
			"IngredientInput": {
				"properties": {
					"name": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unit": {
						"type": "string"
					},
					"notes": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StepInput": {
				"properties": {
					"index": {
						"type": "number",
						"format": "double"
					},
					"instruction": {
						"type": "string"
					},
					"durationMinutes": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"instruction"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SourceType": {
				"type": "string",
				"enum": [
					"ai",
					"manual",
					"import"
				]
			},
			"RecipeStatus": {
				"type": "string",
				"enum": [
					"draft",
					"published"
				]
			},
			"RecipeDTO": {
				"properties": {
					"Id": {
						"type": "string"
					},
					"recipeBookId": {
						"type": "string"
					},
					"originalRecipeId": {
						"type": "string",
						"nullable": true
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"creatorName": {
						"type": "string"
					},
					"categories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"prepTimeMinutes": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"cookTimeMinutes": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"totalTimeMinutes": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"difficulty": {
						"$ref": "#/components/schemas/Difficulty"
					},
					"ingredients": {
						"items": {
							"$ref": "#/components/schemas/IngredientInput"
						},
						"type": "array"
					},
					"steps": {
						"items": {
							"$ref": "#/components/schemas/StepInput"
						},
						"type": "array"
					},
					"notes": {
						"type": "string",
						"nullable": true
					},
					"coverImageUrl": {
						"type": "string",
						"nullable": true
					},
					"sourceType": {
						"$ref": "#/components/schemas/SourceType"
					},
					"sourceId": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/RecipeStatus"
					},
					"createdBy": {
						"type": "string"
					},
					"likes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"title",
					"ingredients",
					"steps"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_RecipeDTO_": {
				"properties": {
					"Id": {
						"type": "string"
					},
					"recipeBookId": {
						"type": "string"
					},
					"originalRecipeId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"creatorName": {
						"type": "string"
					},
					"categories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"prepTimeMinutes": {
						"type": "number",
						"format": "double"
					},
					"cookTimeMinutes": {
						"type": "number",
						"format": "double"
					},
					"totalTimeMinutes": {
						"type": "number",
						"format": "double"
					},
					"difficulty": {
						"$ref": "#/components/schemas/Difficulty"
					},
					"ingredients": {
						"items": {
							"$ref": "#/components/schemas/IngredientInput"
						},
						"type": "array"
					},
					"steps": {
						"items": {
							"$ref": "#/components/schemas/StepInput"
						},
						"type": "array"
					},
					"notes": {
						"type": "string"
					},
					"coverImageUrl": {
						"type": "string"
					},
					"sourceType": {
						"$ref": "#/components/schemas/SourceType"
					},
					"sourceId": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecipeStatus"
					},
					"createdBy": {
						"type": "string"
					},
					"likes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"CommentDTO": {
				"description": "The full comment object returned by the API.",
				"properties": {
					"recipeId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"recipeId",
					"content",
					"createdBy",
					"id",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCommentRequest": {
				"description": "Request body for creating a comment.\r\nTSOA uses this to build the POST Swagger model.",
				"properties": {
					"recipeId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					}
				},
				"required": [
					"recipeId",
					"content",
					"createdBy"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCommentDTO": {
				"description": "Request body for updating a comment.\r\nOnly 'content' is allowed here.",
				"properties": {
					"content": {
						"type": "string"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/SafeUser"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"token"
				],
				"type": "object"
			},
			"RegisterDTO": {
				"properties": {
					"language": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"password",
					"email"
				],
				"type": "object"
			},
			"LoginDTO": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"email"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "assignments",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/createUser": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SafeUser"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserDTO"
							}
						}
					}
				}
			}
		},
		"/users/getUsers": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SafeUser"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/getUserById/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SafeUser"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/updateUser/{id}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SafeUser"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDTO"
							}
						}
					}
				}
			}
		},
		"/users/deleteUser/{id}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/recipes/{id}/image": {
			"post": {
				"operationId": "UploadImage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"url": {
											"type": "string"
										}
									},
									"required": [
										"url"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/recipes/{id}/like": {
			"put": {
				"operationId": "ToggleLike",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"liked": {
											"type": "boolean"
										}
									},
									"required": [
										"liked"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/recipes/ai-search": {
			"get": {
				"operationId": "AiSearch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RecipeDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "recipeBookId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/recipes/createRecipe": {
			"post": {
				"operationId": "CreateRecipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RecipeDTO"
							}
						}
					}
				}
			}
		},
		"/recipes/getRecipeById/{id}": {
			"get": {
				"operationId": "GetRecipeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/recipes/getRecipes": {
			"get": {
				"operationId": "GetRecipes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/RecipeDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "recipeBookId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "difficulty",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/recipes/updateRecipe/{id}": {
			"put": {
				"operationId": "UpdateRecipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecipeDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_RecipeDTO_"
							}
						}
					}
				}
			}
		},
		"/recipes/deleteRecipe/{id}": {
			"delete": {
				"operationId": "DeleteRecipe",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Recipes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/comments/create": {
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentRequest"
							}
						}
					}
				}
			}
		},
		"/comments/getCommentById/{commentId}": {
			"get": {
				"operationId": "GetCommentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "commentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/comments/recipe/getCommentsByRecipe/{recipeId}": {
			"get": {
				"operationId": "GetCommentsByRecipe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CommentDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "recipeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/comments/update/{commentId}": {
			"put": {
				"operationId": "UpdateComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "commentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCommentDTO"
							}
						}
					}
				}
			}
		},
		"/comments/delete/{commentId}": {
			"delete": {
				"operationId": "DeleteComment",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Comments"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "commentId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterDTO"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDTO"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"204": {
						"description": "No Content"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}